contain responses in `ResponseClass`
Rather than making `ResponseClass` a field of the various response
types and all sibling fields Optional, make the various response types
contained in the `Success` variant of `ResponseClass`, and error
information contained in the `Error` variant. This allows the type
system to convey the expected fields, preventing the response types
from eliding the required error fields and failing to parse.
